+++ Makefile
@@ -4,7 +4,7 @@ GIT_COMMIT := $(shell git rev-parse --short HEAD 2>/dev/null || echo "unknown")
 VERSION := $(shell git describe --tags --always --dirty 2>/dev/null || echo "dev")
 
 # Build flags to inject version information
-LDFLAGS := -X main.Version=$(VERSION) -X main.GitCommit=$(GIT_COMMIT) -X main.BuildTime=$(BUILD_DATE) -w -s
+LDFLAGS := -X github.com/Azure/AKSFlexNode/pkg/cmd/version.Version=$(VERSION) -X github.com/Azure/AKSFlexNode/pkg/cmd/version.GitCommit=$(GIT_COMMIT) -X github.com/Azure/AKSFlexNode/pkg/cmd/version.BuildTime=$(BUILD_DATE) -w -s
 
 # Default build for current platform
 .PHONY: build
